Revert "Use psql float instead of bigint to avoid overflow"
This reverts commit fb4825c0464c3e8ed0cf4741536edb9bd3c89e48.

+++ prismaDmob/sql/getClientAllocatorDistributionWeeklyAcc.sql
@@ -11,11 +11,11 @@ with "distributions" as (select date_trunc('week', to_timestamp("height" * 30 + 
                               to_timestamp(3847920 * 30 + 1598306400),
                               current_timestamp,
                               '1 week'::interval) "dates")
-select "weeks"."week"                   as "week",
-       "client"                         as "client",
-       "allocator"                      as "allocator",
-       sum("num_of_allocations")::int   as "num_of_allocations",
-       sum("sum_of_allocations")::float as "sum_of_allocations"
+select "weeks"."week"                    as "week",
+       "client"                          as "client",
+       "allocator"                       as "allocator",
+       sum("num_of_allocations")::int    as "num_of_allocations",
+       sum("sum_of_allocations")::bigint as "sum_of_allocations"
 from "weeks"
          inner join "distributions"
                     on "weeks"."week" >= "distributions"."week"
