+++ prismaDmob/sql/getClientReplicaDistribution.sql
@@ -7,9 +7,9 @@ with "replicas" as (select 'f0' || "clientId"                   as "client",
                     where "termStart" >= 3847920                                           -- nv22 start
                       and to_timestamp("termStart" * 30 + 1598306400) <= current_timestamp -- deals that didn't start yet
                     group by "client", "piece_cid")
-select "client"                      as "client",
-       "num_of_replicas"::int        as "num_of_replicas",
-       sum("total_deal_size")::float as "total_deal_size",
-       sum("piece_size")::float      as "unique_data_size"
+select "client"                       as "client",
+       "num_of_replicas"::int         as "num_of_replicas",
+       sum("total_deal_size")::bigint as "total_deal_size",
+       sum("piece_size")::bigint      as "unique_data_size"
 from "replicas"
 group by "client", "num_of_replicas";
